Set default fps_max on Web to 0.
WaitForTargetFPS() takes a lot of time even when below target FPS, so for best performance, we remove FPS limit.

But, with value of 0, note that Browser will lock FPS to screen refresh rate.
The only way to go above screen refresh rate, is to set FPS limiter to a higher value.